requirements.txt
mem0ai
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
pyyaml
qdrant-client
sentence-transformers
openai